requirements.txt
python-telegram-bot==13.15
python-dotenv==0.21.1
lxml==4.9.1
requests==2.28.2
fake-useragent==1.1.1
beautifulsoup4==4.11.1